+++ functions/cold-transfer.js
@@ -1,0 +1,3 @@
+const { coldTransfer } = require("@spoke-ph/directory-flex-functions");
+
+exports.handler = coldTransfer;