+++ functions/get-directory.js
@@ -1,0 +1,3 @@
+const { getDirectory } = require("@spoke-ph/directory-flex-functions");
+
+exports.handler = getDirectory;